Code changes done for App crashes when I try to increase rows per page in search dashboard issue
