+++ src/dataEntryApp/views/search/SubjectSearch.js
@@ -23,6 +23,7 @@ import { EnhancedTableHead, stableSort, getComparator } from "../../components/T
 import { TablePaginationActions } from "./SubjectSearchPagination";
 import { useTranslation } from "react-i18next";
 import { ToolTipContainer } from "../../components/ToolTipContainer";
+import { SubjectsTable } from "./SubjectSearchTable";
 
 const useStyle = makeStyles(theme => ({
   root: {
@@ -31,9 +32,6 @@ const useStyle = makeStyles(theme => ({
     overflowX: "auto",
     flexShrink: 0,
     marginLeft: theme.spacing(2.5)
-  },
-  table: {
-    minWidth: 1000
   },
   searchCreateToolbar: {
     display: "flex"
@@ -59,226 +57,8 @@ const useStyle = makeStyles(theme => ({
   searchBox: {
     padding: "1.5rem",
     margin: "2rem 1rem"
-  },
-  cellpadding: {
-    padding: "14px 40px 14px 0px"
   }
 }));
-
-const SubjectsTable = ({
-  type,
-  subjects,
-  pageDetails,
-  searchparam,
-  rowsPerPage,
-  setRowsPerPage
-}) => {
-  const classes = useStyle();
-  const { t } = useTranslation();
-  const [order, setOrder] = React.useState("asc");
-  const [orderBy, setOrderBy] = React.useState("fullName");
-  const [selected, setSelected] = React.useState([]);
-  const [page, setPage] = React.useState(0);
-  const [dense, setDense] = React.useState(false);
-
-  let tableHeaderName = [];
-  let pageinfo = pageDetails.subjects;
-  let searchText = searchparam;
-
-  const camelize = str => {
-    return (" " + str).toLowerCase().replace(/[^a-zA-Z0-9]+(.)/g, function(match, chr) {
-      return chr.toUpperCase();
-    });
-  };
-
-  let subjectsListObj = [];
-
-  if (subjects) {
-    subjectsListObj = subjects.map(a => {
-      let firstName = a.firstName ? camelize(a.firstName) : "";
-      let lastName = a.lastName ? camelize(a.lastName) : "";
-      // let subjectType = a.subjectType.name;
-      let sub = {
-        uuid: a.uuid,
-        fullName: firstName + " " + lastName,
-        subjectType: a.subjectType.name,
-        gender: a.gender ? t(a.gender.name) : "",
-        dateOfBirth:
-          new Date().getFullYear() - new Date(a.dateOfBirth).getFullYear() + " " + `${t("years")}`,
-        addressLevel: a.addressLevel ? a.addressLevel.titleLineage : "",
-        activePrograms: a.activePrograms ? a.activePrograms : []
-      };
-      return sub;
-    });
-  }
-
-  if (type.name === "Individual") {
-    tableHeaderName = [
-      { id: "fullName", numeric: false, disablePadding: true, label: "Name", align: "left" },
-      {
-        id: "subjectType",
-        numeric: false,
-        disablePadding: true,
-        label: "subjectType",
-        align: "left"
-      },
-      { id: "gender", numeric: false, disablePadding: true, label: "gender", align: "left" },
-      {
-        id: "dateOfBirth",
-        numeric: true,
-        disablePadding: false,
-        label: "age",
-        align: "left"
-      },
-      {
-        id: "addressLevel",
-        numeric: false,
-        disablePadding: true,
-        label: "addressVillage",
-        align: "left"
-      },
-      {
-        id: "activePrograms",
-        numeric: false,
-        disablePadding: true,
-        label: "enrolments",
-        align: "left"
-      }
-    ];
-  } else {
-    tableHeaderName = [
-      { id: "fullName", numeric: false, disablePadding: true, label: "Name", align: "left" },
-      {
-        id: "addressLevel",
-        numeric: false,
-        disablePadding: true,
-        label: "location",
-        align: "left"
-      },
-      {
-        id: "activePrograms",
-        numeric: false,
-        disablePadding: true,
-        label: "activeprograms",
-        align: "left"
-      }
-    ];
-  }
-
-  const handleRequestSort = (event, property) => {
-    const isAsc = orderBy === property && order === "asc";
-    setOrder(isAsc ? "desc" : "asc");
-    setOrderBy(property);
-  };
-
-  const handleSelectAllClick = event => {
-    if (event.target.checked) {
-      const newSelecteds = subjects.map(n => n.name);
-      setSelected(newSelecteds);
-      return;
-    }
-    setSelected([]);
-  };
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-    // pageDetails.setSearchParams();
-    pageDetails.search({ page: newPage, query: searchText, size: rowsPerPage });
-  };
-
-  const handleChangeRowsPerPage = event => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-    // pageDetails.setSearchParams({ page: 0, query: searchText, size: event.target.value });
-    pageDetails.search({ page: 0, query: searchText, size: rowsPerPage });
-  };
-
-  const isSelected = name => selected.indexOf(name) !== -1;
-
-  // const emptyRows = rowsPerPage - Math.min(rowsPerPage, subjectsListObj.length - rowsPerPage);
-
-  return (
-    subjectsListObj && (
-      <div>
-        <Table className={classes.table} aria-label="custom pagination table">
-          <EnhancedTableHead
-            headername={tableHeaderName}
-            classes={classes}
-            numSelected={selected.length}
-            order={order}
-            orderBy={orderBy}
-            onSelectAllClick={handleSelectAllClick}
-            onRequestSort={handleRequestSort}
-            rowCount={subjectsListObj.length}
-          />
-          <TableBody>
-            {subjectsListObj.map((row, index) => {
-              return (
-                <TableRow key={row.fullName}>
-                  <TableCell component="th" scope="row" padding="none" width="20%">
-                    <Link to={`/app/subject?uuid=${row.uuid}`}>{row.fullName}</Link>
-                  </TableCell>
-                  <TableCell component="th" scope="row" padding="none" width="12%">
-                    {row.subjectType}
-                  </TableCell>
-                  {type.name === "Individual" && (
-                    <TableCell align="left" className={classes.cellpadding}>
-                      {row.gender}
-                    </TableCell>
-                  )}
-                  {type.name === "Individual" && (
-                    <TableCell align="left" className={classes.cellpadding}>
-                      {row.dateOfBirth}
-                    </TableCell>
-                  )}
-                  <TableCell align="left" className={classes.cellpadding}>
-                    {row.addressLevel}
-                  </TableCell>
-                  <TableCell align="left" width="25%" className={classes.cellpadding}>
-                    {" "}
-                    {row.activePrograms.map((p, key) => (
-                      <Button
-                        key={key}
-                        size="small"
-                        style={{
-                          height: 20,
-                          padding: 2,
-                          margin: 2,
-                          backgroundColor: p.colour,
-                          color: "white",
-                          fontSize: 11
-                        }}
-                        disabled
-                      >
-                        {t(p.operationalProgramName)}
-                      </Button>
-                    ))}
-                  </TableCell>
-                </TableRow>
-              );
-            })}
-          </TableBody>
-          <TableFooter>
-            <TableRow>
-              <TablePagination
-                rowsPerPageOptions={[10, 20, 50, { label: "All", value: -1 }]}
-                // component="div"
-                // colSpan={3}
-                search={searchText}
-                count={pageinfo.totalElements}
-                rowsPerPage={rowsPerPage}
-                page={page}
-                onChangePage={handleChangePage}
-                onChangeRowsPerPage={handleChangeRowsPerPage}
-                ActionsComponent={TablePaginationActions}
-              />
-            </TableRow>
-          </TableFooter>
-        </Table>
-      </div>
-    )
-  );
-};
 
 const SubjectSearch = props => {
   const classes = useStyle();
@@ -288,8 +68,8 @@ const SubjectSearch = props => {
 
   const handleSubmit = event => {
     event.preventDefault();
-    console.log("-------------->", event.target);
-    props.search({ query: searchvalue });
+    // console.log("-------------->", event.target);
+    props.search({ page: 0, query: searchvalue, size: rowsPerPage });
   };
 
   const valueSubmit = e => {
@@ -301,7 +81,7 @@ const SubjectSearch = props => {
 
   useEffect(() => {
     props.search();
-    // sessionStorage.clear("subject");
+    sessionStorage.clear("subject");
   }, []);
 
   return (
@@ -343,6 +123,12 @@ const SubjectSearch = props => {
           rowsPerPage={rowsPerPage}
           setRowsPerPage={setRowsPerPage}
         />
+        {/* <SubjectsTable
+          subjects={props.subjects.content}
+          type={props.subjectType}
+          pageDetails={props}
+          searchparam={searchvalue}         
+        /> */}
       </Paper>
     )
   );
